computeSimpleReimbursementPlan from classs to method
